+++ src/test/java/domain/DealerTest.java
@@ -31,25 +31,6 @@ public class DealerTest {
             "9하트,10하트,9,10,DRAW",
             "A하트,10하트,11,10,LOSE"
     })
-    void 플레이어보다_21에_가깝다면_승리(String card1, String card2, int cardValue1, int cardValue2, Result expected) {
-        List<Card> dealerCardsByCardBox = new ArrayList<>();
-        dealerCardsByCardBox.add(new Card("9하트", 9));
-        dealerCardsByCardBox.add(new Card("10하트", 10));
-        Cards dealerCards = new Cards(dealerCardsByCardBox);
-
-        Dealer dealer = new Dealer(dealerCards);
-
-        Name playerName = new Name("wuga");
-        List<Card> playerCardsByCardBox = new ArrayList<>();
-        playerCardsByCardBox.add(new Card(card1, cardValue1));
-        playerCardsByCardBox.add(new Card(card2, cardValue2));
-        Cards playerCards = new Cards(playerCardsByCardBox);
-        Betting betting = new Betting(10000,0);
-
-        Player player = new Player(playerCards, playerName,betting);
-
-        assertThat(dealer.checkWinningResult(dealer,player)).isEqualTo(expected);
-    }
 
     @Test
      void 딜러의_카드총합이_16이하인지_판단(){
